Updated To Work With .Net 3.5
Modified the project (not the solution) as well as required code to work
with .Net framework 3.5.  Any functions with optional parameters are now
overloaded (in some cases renamed to [originalName]With[ParameterName]).
Dynamic keyword in event parsing was removed and replaced with object and
making use of template type.  Implementation of event listener requires
that you check the StripeEvent.Type and declare the type of Stripe object
that will be returned when calling ParseEventDataItem on any event data objects.